requirements.txt
htmltemplate
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
altair<5
beautifulsoup4
huggingface_hub
sentence_transformers
InstructorEmbedding
